requirements.txt
discord.py==2.0.0
asyncio==3.4.3
python-dotenv==0.20.0
svglib==1.4.1
python-dateutil==2.8.2